requirements.txt
pandas==2.3.1
numpy==2.3.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
plotly==6.3.0
streamlit==1.49.1